# frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/scout/apply": {
      "post": {
        "operationId": "ScoutController_apply",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyScoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Scout"
        ]
      }
    },
    "/scout/applications": {
      "get": {
        "operationId": "ScoutController_getMyApplications",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": true,
              "default": 1,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": true,
              "default": 10,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Scout"
        ]
      }
    },
    "/scout/submit": {
      "post": {
        "operationId": "ScoutController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoutSubmitStartupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Scout"
        ]
      }
    },
    "/scout/submissions": {
      "get": {
        "operationId": "ScoutController_getMySubmissions",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": true,
              "default": 1,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": true,
              "default": 10,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "investorId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Scout"
        ]
      }
    },
    "/investor/scout-applications": {
      "get": {
        "operationId": "ScoutController_getScoutApplications",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": true,
              "default": 1,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": true,
              "default": 10,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Scout"
        ]
      }
    },
    "/investor/scout-applications/{id}/approve": {
      "post": {
        "operationId": "ScoutController_approveApplication",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Scout"
        ]
      }
    },
    "/investor/scout-applications/{id}/reject": {
      "post": {
        "operationId": "ScoutController_rejectApplication",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectScoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Scout"
        ]
      }
    },
    "/investor/scout-submissions": {
      "get": {
        "operationId": "ScoutController_getScoutSubmissions",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": true,
              "default": 1,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": true,
              "default": 10,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "investorId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Scout"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS Boilerplate API",
    "description": "Production-ready NestJS boilerplate with auth, queues, storage, and email",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
...
      "ApplyScoutDto": {
        "type": "object",
        "properties": {
          "investorId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "linkedinUrl": {
            "type": "string",
            "format": "uri"
          },
          "experience": {
            "type": "string",
            "minLength": 100,
            "maxLength": 1000
          },
          "motivation": {
            "type": "string",
            "minLength": 100,
            "maxLength": 1000
          },
          "dealflowSources": {
            "type": "string",
            "minLength": 50,
            "maxLength": 500
          },
          "portfolio": {
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "investorId",
          "name",
          "email",
          "linkedinUrl",
          "experience",
          "motivation",
          "dealflowSources"
        ]
      },
      "ScoutSubmitStartupDto": {
        "type": "object",
        "properties": {
          "investorId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "startupData": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 200
              },
              "tagline": {
                "type": "string",
                "minLength": 1,
                "maxLength": 500
              },
              "description": {
                "type": "string",
                "minLength": 100,
                "maxLength": 5000
              },
              "website": {
                "type": "string",
                "format": "uri"
              },
              "location": {
                "type": "string",
                "minLength": 1,
                "maxLength": 200
              },
              "industry": {
                "type": "string",
                "minLength": 1,
                "maxLength": 200
              },
              "stage": {
                "type": "string",
                "enum": [
                  "pre_seed",
                  "seed",
                  "series_a",
                  "series_b",
                  "series_c",
                  "series_d",
                  "series_e",
                  "series_f_plus"
                ]
              },
              "fundingTarget": {
                "type": "integer",
                "exclusiveMinimum": true,
                "maximum": 9007199254740991,
                "minimum": 0
              },
              "teamSize": {
                "type": "integer",
                "exclusiveMinimum": true,
                "maximum": 9007199254740991,
                "minimum": 0
              },
              "pitchDeckUrl": {
                "type": "string",
                "format": "uri"
              },
              "demoUrl": {
                "type": "string",
                "format": "uri"
              }
            },
            "required": [
              "name",
              "tagline",
              "description",
              "website",
              "location",
              "industry",
              "stage",
              "fundingTarget",
              "teamSize"
            ]
          },
          "notes": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "investorId",
          "startupData"
        ]
      },
      "RejectScoutDto": {
        "type": "object",
        "properties": {
          "rejectionReason": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500
          }
        },
        "required": [
          "rejectionReason"
        ]
      },
...
    }
  }
}